feat(point): add an intensity range for point clouds
It is possible to specify the min and the max of the intensity range,
while updating the material in the same time to reflect those changes.

+++ src/Parser/PotreeBinParser.js
@@ -18,7 +18,7 @@ const POINT_ATTTRIBUTES = {
         numElements: 1,
         numByte: 2,
         // using Float32Array because Float16Array doesn't exist
-        arrayType: Float32Array,
+        arrayType: Uint16Array,
         attributeName: 'intensity',
         normalized: true,
     },
